Commit bordel

+++ app/Controller/EventsController.php
@@ -7,12 +7,20 @@
 class EventsController extends AppController {
     public $helpers = array('Html', 'Form', 'Session');
     public $components = array('Session');
-    var $uses = array("Event", "EventLocation","Location");
+    var $uses = array("Event", "EventLocation","Location", "EventImportance", "Setting","Server","Playlist");
     
+    /**
+     * Index page -> List of events
+     */
     public function index() {
         $this->set('events', $this->Event->find('all'));
     }
     
+    /**
+     * View an event
+     * @param type $id
+     * @throws NotFoundException
+     */
     public function view($id = null) {
         if (!$id) {
             throw new NotFoundException(__('Invalid event'));
@@ -24,6 +32,17 @@ class EventsController extends AppController {
             throw new NotFoundException(__('Invalid event'));
         }
         $this->set('event', $event);
+        if (sizeof($event["LiveSession"])>0) {
+            $this->set("displaymonitoring",true);
+        } else {
+            $this->set("displaymonitoring", false);
+        }
+        // Get event importance list
+        $eventImportances = $this->EventImportance->find('list');
+        $this->set(compact('eventImportances'));
+        $this->EventImportance->recursive=-1;
+        $eventImportancesArray = $this->EventImportance->find('all');
+        $this->set('eventImportancesArray', $eventImportancesArray);
         // Get locations
         $locs = $this->Location->find('threaded', array(
                                         'fields' => array('id', 'parent_id', 'name'),
@@ -38,6 +57,14 @@ class EventsController extends AppController {
         $locations=array();
         $this->buildEventLocation($locs, &$locations, "", $eventlocs);
         $this->set(compact('locations'));
+        
+        // Get settings data (to calculate max video bitrate or count of users
+        $setBwRatio = $this->Setting->find('first', array('conditions' => array('Setting.key'=> "bandwidth_threshold_ratio")));
+        $this->set("bandwidth_threshold_ratio",$setBwRatio["Setting"]["value"]);
+        $setBwUsageRatio = $this->Setting->find('first', array('conditions' => array('Setting.key'=> "bandwidth_usage_ratio")));
+        $this->set("bandwidth_usage_ratio",$setBwUsageRatio["Setting"]["value"]);
+        
+        
     }
     
     public function add() {
@@ -72,6 +99,15 @@ class EventsController extends AppController {
             }
             $this->Session->setFlash(__('Unable to add your event.'));
         }
+        
+        // Get event importance list
+        $eventImportances = $this->EventImportance->find('list');
+        $this->set(compact('eventImportances'));
+        $this->EventImportance->recursive=-1;
+        $eventImportancesArray = $this->EventImportance->find('all');
+        $this->set('eventImportancesArray', $eventImportancesArray);
+        
+        
         // Get locations
         $locs = $this->Location->find('threaded', array(
                                         'fields' => array('id', 'parent_id', 'name'),
@@ -86,8 +122,24 @@ class EventsController extends AppController {
         $locations=array();
         $this->buildEventLocation($locs, &$locations, "", $eventlocs);
         $this->set(compact('locations'));
+        
+        // Get settings data (to calculate max video bitrate or count of users
+        $setBwRatio = $this->Setting->find('first', array('conditions' => array('Setting.key'=> "bandwidth_threshold_ratio")));
+        $this->set("bandwidth_threshold_ratio",$setBwRatio["Setting"]["value"]);
+        $setBwUsageRatio = $this->Setting->find('first', array('conditions' => array('Setting.key'=> "bandwidth_usage_ratio")));
+        $this->set("bandwidth_usage_ratio",$setBwUsageRatio["Setting"]["value"]);
+        
+        // Get playlists
+        $playlists = $this->Playlist->find('list');
+        $this->set(compact('playlists'));
     }
     
+    /**
+     * Edit an event
+     * @param type $id
+     * @return type
+     * @throws NotFoundException
+     */
     public function edit($id = null) {
         if (!$id) {
             throw new NotFoundException(__('Invalid event'));
@@ -96,6 +148,9 @@ class EventsController extends AppController {
         $event = $this->Event->findById($id);
         if (!$event) {
             throw new NotFoundException(__('Invalid event'));
+        }
+        if ($event["Event"]["live"]==Configure::read('PUBLISHING_POINT_LIVE_ON')) {
+            throw new ForbiddenException(__('Unable to edit an event in Live state'));
         }
 
         if ($this->request->is(array('post', 'put'))) {
@@ -125,7 +180,7 @@ class EventsController extends AppController {
                 }
                 
                 $this->Session->setFlash(__('Your event has been updated.'));
-                return $this->redirect(array('action' => 'index'));
+                return $this->redirect(array('action' => 'view', $id));
             }
             $this->Session->setFlash(__('Unable to update your event.'));
         }
@@ -134,22 +189,43 @@ class EventsController extends AppController {
             $this->request->data = $event;
         }
         $this->set('event', $event);
+        
+        // Get event importance list
+        $eventImportances = $this->EventImportance->find('list');
+        $this->set(compact('eventImportances'));
+        $this->EventImportance->recursive=-1;
+        $eventImportancesArray = $this->EventImportance->find('all');
+        $this->set('eventImportancesArray', $eventImportancesArray);
+        
         // Get locations
         $locs = $this->Location->find('threaded', array(
                                         'fields' => array('id', 'parent_id', 'name'),
                                         'order' => array('lft ASC') // or array('id ASC')
                                     ));
-        
+       
         // Get EventLocation
         $eventlocs = $this->EventLocation->find('all', array(
                                         'fields' => array('id', 'location_id', 'audience'),
                                         'conditions' => array('EventLocation.event_id' => $id)
                                     ));
         
+        
+        
         $locations=array();
         $this->buildEventLocation($locs, &$locations, "", $eventlocs);
+        
         $this->set(compact('locations'));
         
+        
+        // Get settings data (to calculate max video bitrate or count of users
+        $setBwRatio = $this->Setting->find('first', array('conditions' => array('Setting.key'=> "bandwidth_threshold_ratio")));
+        $this->set("bandwidth_threshold_ratio",$setBwRatio["Setting"]["value"]);
+        $setBwUsageRatio = $this->Setting->find('first', array('conditions' => array('Setting.key'=> "bandwidth_usage_ratio")));
+        $this->set("bandwidth_usage_ratio",$setBwUsageRatio["Setting"]["value"]);
+        
+        // Get playlists
+        $playlists = $this->Playlist->find('list');
+        $this->set(compact('playlists'));
     }
     
     /**
@@ -180,6 +256,9 @@ class EventsController extends AppController {
      * @param type $eventlocs (array of EventLocation for this event)
      */
     private function buildEventLocation($locs, $results, $prefix, $eventlocs) {
+        // Get settings data (to calculate max video bitrate or count of users
+        $setBwRemote = $this->Setting->find('first', array('conditions' => array('Setting.key'=> "remote_bandwidth")));
+        
         foreach ($locs as $loc) {
             $location=array();
             $location["locationname"]=$prefix . $loc["Location"]["name"];
@@ -193,6 +272,23 @@ class EventsController extends AppController {
                 $results[]=$location;
                 $this->buildEventLocation($loc["children"], &$results, $prefix . "&nbsp;&nbsp;&nbsp;", $eventlocs);
             }  else {
+                $bw=0;
+                if (sizeof($loc["Server"])>0) {
+                    // Sum bandwidth for all available servers
+                    foreach ($loc["Server"] as $server) {
+                        $bw+=$server["bandwidth"];
+                    }
+                } else {
+                    // Case of a remote location
+                    // Get list of server for the current location (in remote location)
+                    $serverList = $this->Server->findByLocationId("remote",$loc["Location"]["id"]);
+                    if (sizeof($serverList)>0) {
+                        // If there is at least one server to handle this remote location, 
+                        // we set the bandwidth limit according to settings table
+                        $bw=$setBwRemote["Setting"]["value"];
+                    } 
+                }
+                $location["bandwidth"]=$bw;
                 $results[]=$location;
             }
             
@@ -210,6 +306,10 @@ class EventsController extends AppController {
         if (!$event) {
             throw new NotFoundException(__('Invalid event'));
         }
+        if ($event["Event"]["live"]==Configure::read('PUBLISHING_POINT_LIVE_ON')) {
+            throw new ForbiddenException(__('Unable to delete an event in Live state'));
+        }
+        
         $eventname=$event["Event"]["name"];
         $this->Event->delete($id);
         $this->Session->setFlash(__('The event %s has been deleted.', $eventname));
